okItsWorking

+++ IHaveNoIdeaWhatImDoingAnyMore/User.cs
@@ -11,11 +11,10 @@ namespace IHaveNoIdeaWhatImDoingAnyMore
 {
     using System;
     using System.Collections.Generic;
-
-
-
+    
     public partial class User
     {
+
         public User(User user)
         {
             user_id = user.user_id;
@@ -27,8 +26,7 @@ namespace IHaveNoIdeaWhatImDoingAnyMore
         {
 
         }
-
-        public long user_id { get; set; }
+        public int user_id { get; set; }
         public string username { get; set; }
         public string password { get; set; }
         public int user_access_rank { get; set; }
